requirements.txt
textual~=0.68.0
python-socketio~=5.11.2
websocket-client~=1.8.0
requests~=2.32.3
